moved functions to utilities, fixed relative paths.
