+++ src/generate_tasks.py
@@ -8,10 +8,10 @@ from typing import Any, Dict, List, Tuple
 from langsmith import tracing_context
 from tenacity import Retrying, stop_after_attempt
 
-from capability import Capability, CapabilityState
-from model import Model
-from utils import constants, prompts
-from utils.capability_utils import extract_and_parse_response
+from src.capability import Capability, CapabilityState
+from src.model import Model
+from src.utils import constants, prompts
+from src.utils.capability_utils import extract_and_parse_response
 
 
 logger = logging.getLogger(__name__)
